Fix: Migration Feat: Create Business

+++ database/migrations/2023_01_28_080038_create_business_categories_table.php
@@ -15,7 +15,7 @@ class CreateBusinessCategoriesTable extends Migration
     {
         Schema::create('business_categories', function (Blueprint $table) {
             $table->char("business_id", 36);
-            $table->unsignedInteger("category_id");
+            $table->unsignedBigInteger("category_id");
             $table->foreign('business_id')->references('id')->on('businesses');
             $table->foreign('category_id')->references('id')->on('master_categories');
         });
